correctness: fix silent GridSearch no-op, unscored BSP tree, rule-count overflow
Model-wrapper correctness fixes (autonomous review 2026-08-23).

- TribbleClassifier cached a derived `anomaly_params` in __init__; sklearn
  clone/set_params (GridSearchCV/RandomizedSearchCV) rewrites norm_conorm/t_norm/
  t_conorm without touching the cache, so any norm grid search was a silent no-op.
  Resolve on demand via `_anomaly_params()` (matches sibling classes). Verified:
  clone().set_params(norm_conorm="luk") now takes effect.
- BSPFuzzyTreeClassifier.predict returned an object array, so accuracy_score/
  .score raised "mix of multiclass and unknown". Cast to classes_ dtype; the tree
  is now scorable.
- report.py per-label "Total possible rules" used int64 np.prod -> overflow/garbage
  past ~5^30; use dtype=float like the AND-rules line.
- gauss_plot.plot_gaussians: a dropna'd Series indexed by a full-length boolean
  raised IndexingError on any NaN; align y to the dropped index first.
- gauss_plot random top-k baseline used a wrong formula for k>=2; it is exactly
  k / n_classes under uniform random ranking.

NOTE: finding M2 (IT2/GT2 predict rebuilding X from .values by position) was
investigated and REJECTED -- that positional mapping is the intended sklearn
positional-predict contract (callers pass arrays / differently-named frames);
switching to name-based selection broke test_gt2_regressor with a KeyError.

Tests: 49 classifier tests pass (incl. norm_passthrough); it2/gt2 suites green.

+++ src/tribblefis/bsp_fuzzy_classifier.py
@@ -220,7 +220,8 @@ class BSPFuzzyTreeClassifier(BaseEstimator, ClassifierMixin):
         for node in {id(l): l for l in leaves}.values():
             mask = np.array([leaves[i] is node for i in range(len(X_df))])
             preds[mask] = node.model.predict(X_df[mask])
-        return preds
+        # Cast off the object dtype so accuracy_score/score can type the labels.
+        return preds.astype(self.classes_.dtype, copy=False)
 
     @property
     def n_leaves_(self) -> int:
